MA-11839-2 Add M4 image for imx8qm/imx8qxp
Add M4 image support for imx8qm and imx8qxp.

Reviewed-on: http://androidsource.ap.freescale.net/project/4587

+++ mek_8q/AndroidUboot.mk
@@ -5,17 +5,26 @@ define build_uboot
 	if [ "$(strip $(2))" == "imx8qm" ]; then \
 		MKIMAGE_PLATFORM=`echo iMX8QM`; \
 		SCFW_PLATFORM=`echo 8qm`;  \
-		FLASH_TARGET=`echo flash`;  \
 		if [ "$(PRODUCT_IMX_CAR)" != "true" ]; then \
+			FLASH_TARGET=`echo flash`;  \
 			cp $(FSL_PROPRIETARY_PATH)/linux-firmware-imx/firmware/hdmi/cadence/hdmitxfw.bin $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/hdmitxfw.bin; \
-		elif [ -f $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/hdmitxfw.bin ]; then \
-			rm -f $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/hdmitxfw.bin; \
+		else \
+			FLASH_TARGET=`echo flash_multi_cores_m4_1`;  \
+			if [ -f $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/hdmitxfw.bin ]; then \
+				rm -f $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/hdmitxfw.bin; \
+			fi; \
+			cp  $(FSL_PROPRIETARY_PATH)/fsl-proprietary/mcu-sdk/imx8q/imx8qm_m4_1_TCM_hello_world.bin $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/m41_tcm.bin; \
 		fi; \
 	elif [ "$(strip $(2))" == "imx8qxp" ]; then \
 		MKIMAGE_PLATFORM=`echo iMX8QX`; \
 		SCFW_PLATFORM=`echo 8qx`; \
-		FLASH_TARGET=`echo flash_b0`;  \
 	        cp  $(FSL_PROPRIETARY_PATH)/linux-firmware-imx/firmware/seco/ahab-container.img $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/ahab-container.img; \
+		if [ "$(PRODUCT_IMX_CAR)" == "true" ]; then \
+			FLASH_TARGET=`echo flash_b0_all`;  \
+			cp  $(FSL_PROPRIETARY_PATH)/fsl-proprietary/mcu-sdk/imx8q/imx8qx_m4_TCM_hello_world.bin $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/CM4.bin; \
+		else \
+			FLASH_TARGET=`echo flash_b0`;  \
+		fi; \
 	fi; \
 	cp  out/target/product/mek_8q/obj/BOOTLOADER_OBJ/u-boot.$(strip $(1)) $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/u-boot.bin; \
 	cp  $(FSL_PROPRIETARY_PATH)/fsl-proprietary/uboot-firmware/imx8q/mx$$SCFW_PLATFORM-scfw-tcm.bin $(IMX_MKIMAGE_PATH)/imx-mkimage/$$MKIMAGE_PLATFORM/scfw_tcm.bin; \
